move compiling to test to own folder after code review
