--- java-checks-test-sources/default/src/main/java/checks/visibility/restriction/README.md
@@ -1,1 +1,0 @@
-Use this package to restrict the visibility of classes for a check sample. For instance S3252, we use it to verify that S3252 doesn't raise an issue when the class owning a static field is not accessible from the current package
